refactor: extract types TensorLike and Number to new tdecomp.types
