+++ tests/test_utils.py
@@ -8,4 +8,4 @@ def test_pseudo_inverse():
     C = torch.rand((3, 3))
     c_inv_torch = torch.linalg.pinv(C)
     c_inv_solve = tdecomp.utils.pseudo_inverse(C)
-    assert torch.allclose(c_inv_solve, c_inv_torch)
+    assert torch.allclose(c_inv_solve, c_inv_torch, atol=1e-5)
